requirements.txt
amqp==2.3.2
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
boto3==1.9.41
botocore==1.12.41
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.1
daphne==1.4.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-appconf==1.0.2
django-bootstrap-forms==0.1
django-bootstrap4==0.0.7
django-braces==1.13.0
django-countries==5.3.2
django-crispy-forms==1.7.0
django-filter==2.0.0
django-filters==0.2.1
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-invitations==1.9.2
django-mailgun==0.9.1
django-model-utils==3.1.2
django-paypal==0.5.0
django-phonenumber-field==2.1.0
django-private-chat==0.2.2
django-separatedvaluesfield==0.6.0
django-storages==1.7.1
django-stubs==0.8.2
django-taggit==0.24.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
djangorestframework-stubs==0.2.0
djoser==1.5.0
docutils==0.14
enum34==1.1.6
Fabric3==1.14.post1
gunicorn==19.7.1
idna==2.7
ipaddress==1.0.17
jmespath==0.9.3
jsonpickle==1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
Markdown==3.0.1
mypy==0.670
mypy-extensions==0.4.1
ndg-httpsclient==0.5.1
oauthlib==2.1.0
paramiko==2.4.2
phonenumbers==8.10.0
Pillow==5.2.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
pusher==2.1.3
pyasn1==0.4.4
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.7.5
python-decouple==3.1
python-digitalocean==1.14.0
python-dotenv==0.10.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.4
pywin32-ctypes==0.2.0
pyxdg==0.25
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
SecretStorage==2.3.1
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.23
vine==1.1.4
websockets==7.0
whitenoise==3.3.1
zope.interface==4.6.0
